fix: break circular import between utils.tools and metadata_extractor.io by moving collect_file_paths to utils.file_listings
